Split parts of XmlBuilder class into sperate files. Changed buid logic of the dictionary componet to create the XElement.
